Attempt at using typeorm
Delete doesnt seem to work

+++ attemptApi/src/database/database-init.ts
@@ -1,32 +1,36 @@
-import { AttemptEntity } from "./attemptEntity";
-import { AttemptDataSource } from "./database";
+import { AttemptEntity } from './attemptEntity';
+import { AttemptDataSource } from './database';
 
-async function run() {
-    await AttemptDataSource.initialize()
-    if (await doEntityExist()) {
-        console.log("Initializing or Recreating tables ... ")
-        deleteTable()
-        await AttemptDataSource.synchronize()
-    } else {
-        console.log("Table is already initialized ")
+export async function run(forceDelete : boolean = false) {
+  await AttemptDataSource.initialize();
+  if (await doEntityExist()) {
+    console.log('Initializing or Recreating tables ... ');
+    if (forceDelete) {
+      await deleteTable();
     }
+    await AttemptDataSource.synchronize();
+  } else {
+    console.log('Table is already initialized ');
+    await AttemptDataSource.synchronize();
+  }
 }
 
 async function doEntityExist() {
-    return (await AttemptDataSource.query(
+  return (
+    (
+      await AttemptDataSource.query(
         'SELECT 1 FROM information_schema.tables WHERE table_name IN ($1)',
         [AttemptDataSource.getRepository(AttemptEntity).metadata.tableName],
-    )).length > 0;
+      )
+    ).length > 0
+  );
 }
 
 async function deleteTable() {
-    await AttemptDataSource.query(
-        `DROP TABLE IF EXISTS ${AttemptDataSource.getRepository(AttemptEntity).metadata.tableName}`,
-    )
-}
+  console.log('deleting table...')
+  const tname: string = AttemptDataSource.getRepository(AttemptEntity).metadata.tableName
 
-console.log('a')
-async function Fa() {
-
-
+  // await AttemptDataSource.
+  await AttemptDataSource.query(
+    `DROP TABLE IF EXISTS ${tname}`,);
 }
